If number is not 1 or allele-specific, check if any value passes
For features not explicitly linked with alleles (i.e. not Number=A or Number=R),
any value meeting expression criteria will be considered as all alleles meeting
the expression criteria. It should therefore be strongly recommended to fix
numbering of values in headers from genotypers that do not respect this
numbering system (e.g. NV in Platypus output should be changed from Number=. to
Number=A).